Create UI for Rock-paper-scissors game
created an UI by creating elements for buttons and results.
then changed the codes of the functions to make sure that everything worked.
challenges in rewriting the original getHumanChoice-function to a function
that reacted on the chosen buttons (by creatin a handleClick-function).
endresult:
- created elements for buttons and results;
- used the old function combination to let the computer make a choice
(and found out about a simpler function alternative);
- used the old function to play a round;
- created a new handleClick-function that used the input from the computer
and the output from the addEventListeners for the three buttons, to produce
a result of both parties and declare a winner, and to add to the total score.